requirements.txt
groq>=0.9.0
tavily-python>=0.7.0
python-dotenv>=1.0.1
rich>=13.9.4
